[api rearrange] project started. mecab_direct_connecter is now moved morphing.
